+++ frontend/src/components/Home.js
@@ -79,8 +79,10 @@ function Home() {
         .then(data => { 
           console.log('From home /api/settings:', data); 
           setWssHost(data.wssHost); 
-          setQrCode(data.qrPairingCode); 
-          setPairingCode(data.manualPairingCode);
+          if(data.matterbridgeInformation.bridgeMode==='bridge') {
+            setQrCode(data.qrPairingCode); 
+            setPairingCode(data.manualPairingCode);
+          }
           setSystemInfo(data.systemInformation);
           setMatterbridgeInfo(data.matterbridgeInformation);
           localStorage.setItem('wssHost', data.wssHost);
@@ -113,15 +115,17 @@ function Home() {
     if (selectedRow === row) {
       setSelectedRow(-1);
       setSelectedPluginName('none');
-      setQrCode(localStorage.getItem('qrPairingCode'));
-      setPairingCode(localStorage.getItem('manualPairingCode'));
+      // setQrCode(localStorage.getItem('qrPairingCode'));
+      // setPairingCode(localStorage.getItem('manualPairingCode'));
+      setQrCode('');
+      setPairingCode('');
     } else {
       setSelectedRow(row);
       setSelectedPluginName(plugins[row].name);
       setQrCode(plugins[row].qrPairingCode);
       setPairingCode(plugins[row].manualPairingCode);
     }
-    console.log('Selected row:', row, 'plugin:', plugins[row].name, 'qrcode:', plugins[row].qrPairingCode);
+    // console.log('Selected row:', row, 'plugin:', plugins[row].name, 'qrcode:', plugins[row].qrPairingCode);
   };
 
   const handleEnableDisable = (row) => {
@@ -206,9 +210,10 @@ function Home() {
         </DialogContent>
       </Dialog>
 
-      <div  style={{ display: 'flex', flexDirection: 'column', height: 'calc(100vh - 60px - 40px)', maxHeight: 'calc(100vh - 60px - 40px)', flex: '0 1 auto', gap: '20px' }}>
-        {qrCode && <QRDiv qrText={qrCode} pairingText={pairingCode} qrWidth={256} topText="QRCode" bottomText={selectedPluginName==='none'?'Matterbridge':selectedPluginName}/>}
+      <div  style={{ display: 'flex', flexDirection: 'column', height: 'calc(100vh - 60px - 40px)', maxHeight: 'calc(100vh - 60px - 40px)', width: '302px', minWidth: '302px', maxWidth: '302px', flex: '1 1 auto', gap: '20px' }}>
+        {qrCode && <QRDiv qrText={qrCode} pairingText={pairingCode} qrWidth={256} topText="QR pairing code" bottomText={selectedPluginName==='none'?'Matterbridge':selectedPluginName} matterbridgeInfo={matterbridgeInfo} plugin={selectedRow===-1?undefined:plugins[selectedRow]}/>}
         {systemInfo && <SystemInfoTable systemInfo={systemInfo} compact={true}/>}
+        {qrCode==='' && matterbridgeInfo && <MatterbridgeInfoTable matterbridgeInfo={matterbridgeInfo}/>}
       </div>
       <div  style={{ display: 'flex', flexDirection: 'column', height: 'calc(100vh - 60px - 40px)', maxHeight: 'calc(100vh - 60px - 40px)', flex: '1 1 auto', gap: '20px' }}>
 
@@ -236,12 +241,12 @@ function Home() {
                   <td className="table-content"><Tooltip title={plugin.path}>{plugin.name}</Tooltip></td>
                   <td className="table-content">{plugin.description}</td>
                   <td className="table-content"><Tooltip title="Update the plugin to the latest version">{plugin.latestVersion === undefined || plugin.latestVersion === plugin.version ? plugin.version : <span className="status-warning" onClick={() => handleUpdate(index)}>{`${plugin.version} -> ${plugin.latestVersion}`}</span>}</Tooltip></td>
-                  <td className="table-content">{plugin.author}</td>
-                  <td className="table-content">{plugin.type}</td>
+                  <td className="table-content">{plugin.author.replace('https://github.com/', '')}</td>
+                  <td className="table-content">{plugin.type === 'DynamicPlatform'?'Dynamic':'Accessory'}</td>
                   <td className="table-content">{plugin.registeredDevices}</td>
                   <td className="table-content">  
                     <>
-                      {plugin.qrPairingCode ? <Tooltip title="Scan the QRCode"><IconButton style={{padding: 0}} className="PluginsIconButton" onClick={() => handleSelectQRCode(index)} size="small"><QrCode2 /></IconButton></Tooltip> : <></>}
+                      {matterbridgeInfo && matterbridgeInfo.bridgeMode === 'childbridge' ? <Tooltip title="Shows the QRCode or the fabrics"><IconButton style={{padding: 0}} className="PluginsIconButton" onClick={() => handleSelectQRCode(index)} size="small"><QrCode2 /></IconButton></Tooltip> : <></>}
                       <Tooltip title="Plugin config"><IconButton style={{padding: 0}} className="PluginsIconButton" onClick={() => handleConfigPlugin(index)} size="small"><Settings /></IconButton></Tooltip>
                       <Tooltip title="Remove the plugin"><IconButton style={{padding: 0}} className="PluginsIconButton" onClick={() => handleRemovePlugin(index)} size="small"><DeleteForever /></IconButton></Tooltip>
                       {plugin.enabled ? <Tooltip title="Disable the plugin"><IconButton style={{padding: 0}} className="PluginsIconButton" onClick={() => handleEnableDisable(index)} size="small"><Unpublished /></IconButton></Tooltip> : <></>}
@@ -435,62 +440,152 @@ function SystemInfoTable({ systemInfo, compact }) {
   }
 
   return (
-    <table>
-      <thead>
-        <tr>
-          <th colSpan="2" className="table-header">System Information</th>
-        </tr>
-      </thead>
-      <tbody>
-        {Object.entries(systemInfo).filter(([key, _]) => !excludeKeys.includes(key)).map(([key, value], index) => (
-          <tr key={key} className={index % 2 === 0 ? 'table-content-even' : 'table-content-odd'}>
-            <td className="table-content">{key}</td>
-            <td className="table-content">{value}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
+    <div className="MbfWindowDiv">
+      <div className="MbfWindowDivTable">
+        <table>
+          <thead>
+            <tr>
+              <th colSpan="2" className="table-header">System Information</th>
+            </tr>
+          </thead>
+          <tbody>
+            {Object.entries(systemInfo).filter(([key, _]) => !excludeKeys.includes(key)).map(([key, value], index) => (
+              <tr key={key} className={index % 2 === 0 ? 'table-content-even' : 'table-content-odd'} style={{ borderTop: '1px solid #ddd' }}>
+                <td className="table-content">{key}</td>
+                <td className="table-content">
+                  <TruncatedText value={typeof value !== 'string' ? value.toString() : value} maxChars={26} />
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      </div>
+    </div>
   );
 }
 
 // This function takes systemInfo as a parameter and returns a table element with the systemInfo
 function MatterbridgeInfoTable({ matterbridgeInfo }) {
+  const excludeKeys = ['matterbridgeVersion', 'matterbridgeLatestVersion', 'debugEnabled', 'bridgeMode', 'matterbridgeFabricInformations', 'matterbridgeSessionInformations'];
+  if(matterbridgeInfo.bridgeMode === 'childbridge') excludeKeys.push('matterbridgePaired', 'matterbridgeConnected');
   return (
-    <table>
-      <thead>
-        <tr>
-          <th colSpan="2" className="table-header">Matterbridge Information</th>
-        </tr>
-      </thead>
-      <tbody>
-        {Object.entries(matterbridgeInfo).map(([key, value], index) => (
-          <tr key={key} className={index % 2 === 0 ? 'table-content-even' : 'table-content-odd'}>
-            <td className="table-content">{key}</td>
-            <td className="table-content">{value}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-}
-
-// This function takes four parameters: qrText, qrWidth, topText, and bottomText
-// It returns a div element with a rectangle, a QR code, and two texts
-function QRDiv({ qrText, pairingText, qrWidth, topText, bottomText }) {
-  return (
-    <div className="MbfWindowDiv" style={{alignItems: 'center'}} minWidth='360px'>
-      <div className="MbfWindowHeader">
-        <p className="MbfWindowHeaderText" style={{textAlign: 'center'}}>{topText}</p>
-      </div>
-      <QRCode value={qrText} size={qrWidth} bgColor='#9e9e9e' style={{ margin: '20px' }}/>
-      <div  className="MbfWindowFooter">
-        <div>
-          <p style={{ margin: 0, textAlign: 'center', fontSize: '14px' }}>Use {pairingText} or scan the QR to pair</p>
-          <p className="text-color-selected" style={{ margin: 0, textAlign: 'center', fontSize: '14px' }}>{bottomText}</p>
-        </div>
+    <div className="MbfWindowDiv">
+      <div className="MbfWindowDivTable">
+        <table>
+          <thead>
+            <tr>
+              <th colSpan="2" className="table-header">Matterbridge Information</th>
+            </tr>
+          </thead>
+          <tbody>
+            {Object.entries(matterbridgeInfo)
+              .filter(([key, value]) => !excludeKeys.includes(key) && value !== undefined && value !== '')
+              .map(([key, value], index) => (
+              <tr key={key} className={index % 2 === 0 ? 'table-content-even' : 'table-content-odd'} style={{ borderTop: '1px solid #ddd' }}>
+                <td className="table-content">{key.replace('matterbridgeConnected', 'connected').replace('matterbridgePaired', 'paired').replace('homeDirectory', 'home').replace('rootDirectory', 'root').replace('matterbridgeDirectory', 'storage').replace('matterbridgePluginDirectory', 'plugins').replace('globalModulesDirectory', 'modules')}</td>
+                <td className="table-content">
+                  <TruncatedText value={typeof value !== 'string' ? value.toString() : value} maxChars={30} />
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
       </div>
     </div>
   );
+}
+
+function TruncatedText({ value, maxChars }) {
+  let displayText = value;
+  if (value.length > maxChars && maxChars > 3) { 
+    const charsToShow = maxChars - 3; 
+    const start = value.substring(0, Math.ceil(charsToShow / 2));
+    const end = value.substring(value.length - Math.floor(charsToShow / 2), value.length);
+    displayText = `${start} … ${end}`;
+  }
+  if(value !== displayText) return (
+    <Tooltip title={value} placement="top" PopperProps={{
+      modifiers: [
+        {
+          name: 'offset',
+          options: {
+            offset: [0, 12], 
+          },
+        },
+      ],
+    }}>
+      <span>{displayText}</span>
+    </Tooltip>
+  )
+  else return <span>{displayText}</span>
+}
+
+function QRDiv({ qrText, pairingText, qrWidth, topText, bottomText, matterbridgeInfo, plugin }) {
+  // console.log('QRDiv:', matterbridgeInfo, plugin);
+  if(matterbridgeInfo.bridgeMode === 'bridge' && matterbridgeInfo.matterbridgePaired === true) 
+    return ( 
+      <div className="MbfWindowDiv" style={{alignItems: 'center', minWidth: '300px', overflow: 'hidden'}} >
+        <div className="MbfWindowHeader">
+          <p className="MbfWindowHeaderText" style={{textAlign: 'left', overflow: 'hidden'}}>Paired fabrics</p>
+        </div>
+        <div className="MbfWindowBodyColumn">
+          {matterbridgeInfo.matterbridgeFabricInformations.map((fabric, index) => (
+            <div key={index} style={{ margin: '0px 15px 0px 15px', padding: '10px', gap: '0px', backgroundColor: '#9e9e9e', textAlign: 'left', fontSize: '14px' }}>
+                <p className="status-blue" style={{ margin: '0px 10px 5px 10px', fontSize: '14px', padding: 0 }}>Fabric: {fabric.fabricIndex}</p>
+                <p style={{ margin: '0px 20px 0px 20px'}}>Vendor: {fabric.rootVendorId} {fabric.rootVendorName}</p>
+                {fabric.label !== '' && <p style={{ margin: '0px 20px 0px 20px'}}>Label: {fabric.label}</p>}
+                <p style={{ margin: '0px 20px 0px 20px'}}>Active sessions: {matterbridgeInfo.matterbridgeSessionInformations.filter(session => session.fabric.fabricIndex === fabric.fabricIndex).length}</p>
+            </div>  
+          ))}
+        </div>  
+      </div>
+  )  
+  else if(matterbridgeInfo.bridgeMode === 'childbridge' && plugin && plugin.paired === true) 
+    return ( 
+      <div className="MbfWindowDiv" style={{alignItems: 'center', minWidth: '300px', overflow: 'hidden'}} >
+        <div className="MbfWindowHeader">
+          <p className="MbfWindowHeaderText" style={{textAlign: 'left'}}>Paired fabrics</p>
+        </div>
+        <div className="MbfWindowBodyColumn">
+          {plugin.fabricInformations.map((fabric, index) => (
+            <div key={index} style={{ margin: '0px 15px 0px 15px', padding: '10px', gap: '0px', backgroundColor: '#9e9e9e', textAlign: 'left', fontSize: '14px' }}>
+                <p className="status-blue" style={{ margin: '0px 10px 5px 10px', fontSize: '14px', padding: 0 }}>Fabric: {fabric.fabricIndex}</p>
+                <p style={{ margin: '0px 20px 0px 20px' }}>Vendor: {fabric.rootVendorId} {fabric.rootVendorName}</p>
+                {fabric.label !== '' && <p style={{ margin: '0px 20px 0px 20px'}}>Label: {fabric.label}</p>}
+                <p style={{ margin: '0px 20px 0px 20px' }}>Active sessions: {plugin.sessionInformations.filter(session => session.fabric.fabricIndex === fabric.fabricIndex).length}</p>
+            </div>  
+          ))}
+        </div>  
+      </div>
+  )
+  else if(matterbridgeInfo.bridgeMode === 'bridge' && matterbridgeInfo.matterbridgePaired !== true) 
+    return (
+      <div className="MbfWindowDiv" style={{alignItems: 'center', minWidth: '300px'}}>
+        <div className="MbfWindowHeader">
+          <p className="MbfWindowHeaderText" style={{textAlign: 'left'}}>{topText}</p>
+        </div>
+        <QRCode value={qrText} size={qrWidth} bgColor='#9e9e9e' style={{ margin: '20px' }}/>
+        <div className="MbfWindowFooter" style={{padding: 0, marginTop: '-5px', height: '30px'}}>
+          <div>
+            <p style={{ margin: 0, textAlign: 'center', fontSize: '14px' }}>Manual pairing code: {pairingText}</p>
+          </div>
+        </div>
+      </div>
+    );
+  else if(matterbridgeInfo.bridgeMode === 'childbridge' && plugin && plugin.paired !== true)  
+    return (
+      <div className="MbfWindowDiv" style={{alignItems: 'center', minWidth: '300px'}}>
+        <div className="MbfWindowHeader">
+          <p className="MbfWindowHeaderText" style={{textAlign: 'left'}}>{topText}</p>
+        </div>
+        <QRCode value={qrText} size={qrWidth} bgColor='#9e9e9e' style={{ margin: '20px' }}/>
+        <div className="MbfWindowFooter" style={{padding: 0, marginTop: '-5px', height: '30px'}}>
+          <div>
+            <p style={{ margin: 0, textAlign: 'center', fontSize: '14px' }}>Manual pairing code: {pairingText}</p>
+          </div>
+        </div>
+      </div>
+    );
 }
 
 function DialogConfigPlugin( { config, schema, handleCloseConfig }) {